+++ doc/release_notes.rst
@@ -15,6 +15,8 @@ Upcoming Release
 
 * Bug fixing (script retrieve_databundle) and rule run_test to ease testing `PR #322 <https://github.com/pypsa-meets-africa/pypsa-africa/pull/322>`__
 
+* Handling non-numerical entries in raw OSM data: `PR #287 <https://github.com/pypsa-meets-africa/pypsa-africa/pull/287>`__
+
 PyPSA-Africa 0.0.2 (6th April 2022)
 =====================================